requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
botasaurus==1.0.0
pyyaml==6.0.1
fastapi==0.104.1
uvicorn==0.24.0
pytest==7.4.3
pytest-cov==4.1.0
python-dateutil==2.8.2
urllib3==2.0.7
lxml==4.9.3
fake-useragent==1.4.0
python-dotenv==1.0.0
